fix: address long-horizon review feedback
