+++ packages/daemon/src/storage/repositories/space-task-repository.ts
@@ -22,10 +22,6 @@ export class SpaceTaskRepository {
 		private reactiveDb?: ReactiveDatabase
 	) {}
 
-	getDatabase(): BunDatabase {
-		return this.db;
-	}
-
 	private hasMessageSearchIndex(): boolean {
 		try {
 			const row = this.db
