+++ packages/daemon/src/storage/schema/index.ts
@@ -9,6 +9,7 @@
 
 import type { Database as BunDatabase } from 'bun:sqlite';
 import { createEvolutionTables } from './evolution';
+import { createLongHorizonAgentTables } from './long-horizon-agents';
 import { DEFAULT_GLOBAL_TOOLS_CONFIG, DEFAULT_GLOBAL_SETTINGS } from '@neokai/shared';
 
 // Re-export migrations
@@ -89,8 +90,6 @@ export { runMigration123 } from './migrations';
 // knip-ignore-next-line
 export { runMigration124 } from './migrations';
 // knip-ignore-next-line
-export { runMigration144 } from './migrations';
-// knip-ignore-next-line
 export { runMigration125 } from './migrations';
 // knip-ignore-next-line
 export { runMigration126 } from './migrations';
@@ -120,6 +119,8 @@ export { configureMessageSearchFts, runMigration141 } from './migrations';
 export { runMigration142 } from './migrations';
 // knip-ignore-next-line
 export { runMigration143 } from './migrations';
+// knip-ignore-next-line
+export { runMigration144 } from './migrations';
 
 /**
  * Create all database tables and initialize defaults
@@ -748,125 +749,6 @@ function createSpaceAgentInboxTables(db: BunDatabase): void {
 	db.exec(
 		`CREATE INDEX IF NOT EXISTS idx_sessions_space_agent_provenance ` +
 			`ON sessions(json_extract(session_context, '$.spaceId'), json_extract(metadata, '$.promptProvenance.agentId'))`
-	);
-}
-
-export function createLongHorizonAgentTables(db: BunDatabase): void {
-	db.exec(`
-		CREATE TABLE IF NOT EXISTS space_long_horizon_agents (
-			id TEXT PRIMARY KEY,
-			space_id TEXT NOT NULL,
-			handle TEXT NOT NULL,
-			display_name TEXT NOT NULL,
-			template_key TEXT DEFAULT NULL,
-			status TEXT NOT NULL DEFAULT 'active'
-				CHECK(status IN ('active', 'paused', 'disabled', 'archived')),
-			session_id TEXT DEFAULT NULL,
-			instructions TEXT NOT NULL DEFAULT '',
-			autonomy_level INTEGER DEFAULT NULL
-				CHECK(autonomy_level IS NULL OR autonomy_level BETWEEN 1 AND 5),
-			tool_permissions_json TEXT NOT NULL DEFAULT '{}',
-			created_at INTEGER NOT NULL,
-			updated_at INTEGER NOT NULL,
-			FOREIGN KEY (space_id) REFERENCES spaces(id) ON DELETE CASCADE
-		)
-	`);
-	db.exec(
-		`CREATE UNIQUE INDEX IF NOT EXISTS idx_space_long_horizon_agents_handle ` +
-			`ON space_long_horizon_agents(space_id, handle) WHERE status != 'archived'`
-	);
-	db.exec(
-		`CREATE INDEX IF NOT EXISTS idx_space_long_horizon_agents_space_status ` +
-			`ON space_long_horizon_agents(space_id, status)`
-	);
-	db.exec(`
-		CREATE TABLE IF NOT EXISTS space_long_horizon_agent_goals (
-			agent_id TEXT NOT NULL,
-			goal_id TEXT NOT NULL,
-			relationship TEXT NOT NULL DEFAULT 'owner'
-				CHECK(relationship IN ('owner', 'manager', 'watcher')),
-			created_at INTEGER NOT NULL,
-			updated_at INTEGER NOT NULL,
-			PRIMARY KEY(agent_id, goal_id, relationship),
-			FOREIGN KEY (agent_id) REFERENCES space_long_horizon_agents(id) ON DELETE CASCADE,
-			FOREIGN KEY (goal_id) REFERENCES space_goals(id) ON DELETE CASCADE
-		)
-	`);
-	db.exec(
-		`CREATE INDEX IF NOT EXISTS idx_space_lh_agent_goals_goal ` +
-			`ON space_long_horizon_agent_goals(goal_id)`
-	);
-	db.exec(`
-		CREATE TABLE IF NOT EXISTS space_long_horizon_agent_forge_scopes (
-			agent_id TEXT NOT NULL,
-			scope_id TEXT NOT NULL,
-			relationship TEXT NOT NULL DEFAULT 'owner'
-				CHECK(relationship IN ('owner', 'manager', 'watcher')),
-			created_at INTEGER NOT NULL,
-			updated_at INTEGER NOT NULL,
-			PRIMARY KEY(agent_id, scope_id, relationship),
-			FOREIGN KEY (agent_id) REFERENCES space_long_horizon_agents(id) ON DELETE CASCADE,
-			FOREIGN KEY (scope_id) REFERENCES evolution_scopes(id) ON DELETE CASCADE
-		)
-	`);
-	db.exec(
-		`CREATE INDEX IF NOT EXISTS idx_space_lh_agent_scopes_scope ` +
-			`ON space_long_horizon_agent_forge_scopes(scope_id)`
-	);
-	db.exec(`
-		CREATE TABLE IF NOT EXISTS space_long_horizon_agent_reminders (
-			id TEXT PRIMARY KEY,
-			space_id TEXT NOT NULL,
-			agent_id TEXT NOT NULL,
-			title TEXT NOT NULL,
-			body TEXT NOT NULL DEFAULT '',
-			status TEXT NOT NULL DEFAULT 'active'
-				CHECK(status IN ('active', 'paused', 'fired', 'cancelled')),
-			trigger_type TEXT NOT NULL CHECK(trigger_type IN ('at', 'cron')),
-			run_at INTEGER DEFAULT NULL,
-			cron_expression TEXT DEFAULT NULL,
-			timezone TEXT NOT NULL DEFAULT 'UTC',
-			next_run_at INTEGER DEFAULT NULL,
-			last_fired_at INTEGER DEFAULT NULL,
-			created_by_session TEXT DEFAULT NULL,
-			created_at INTEGER NOT NULL,
-			updated_at INTEGER NOT NULL,
-			FOREIGN KEY (space_id) REFERENCES spaces(id) ON DELETE CASCADE,
-			FOREIGN KEY (agent_id) REFERENCES space_long_horizon_agents(id) ON DELETE CASCADE
-		)
-	`);
-	db.exec(
-		`CREATE INDEX IF NOT EXISTS idx_space_lh_agent_reminders_due ` +
-			`ON space_long_horizon_agent_reminders(status, next_run_at) WHERE status = 'active'`
-	);
-	db.exec(
-		`CREATE INDEX IF NOT EXISTS idx_space_lh_agent_reminders_agent ` +
-			`ON space_long_horizon_agent_reminders(agent_id, status)`
-	);
-	db.exec(`
-		CREATE TABLE IF NOT EXISTS space_long_horizon_agent_event_subscriptions (
-			id TEXT PRIMARY KEY,
-			space_id TEXT NOT NULL,
-			agent_id TEXT NOT NULL,
-			source TEXT NOT NULL,
-			topic TEXT NOT NULL,
-			filter_json TEXT NOT NULL DEFAULT '{}',
-			status TEXT NOT NULL DEFAULT 'active'
-				CHECK(status IN ('active', 'paused', 'disabled')),
-			created_at INTEGER NOT NULL,
-			updated_at INTEGER NOT NULL,
-			UNIQUE(space_id, agent_id, source, topic, filter_json),
-			FOREIGN KEY (space_id) REFERENCES spaces(id) ON DELETE CASCADE,
-			FOREIGN KEY (agent_id) REFERENCES space_long_horizon_agents(id) ON DELETE CASCADE
-		)
-	`);
-	db.exec(
-		`CREATE INDEX IF NOT EXISTS idx_space_lh_agent_subscriptions_route ` +
-			`ON space_long_horizon_agent_event_subscriptions(space_id, source, topic, status)`
-	);
-	db.exec(
-		`CREATE INDEX IF NOT EXISTS idx_space_lh_agent_subscriptions_agent ` +
-			`ON space_long_horizon_agent_event_subscriptions(agent_id, status)`
 	);
 }
