+++ packages/daemon/tests/unit/helpers/space-test-db.ts
@@ -12,8 +12,8 @@
  */
 
 import type { Database as BunDatabase } from 'bun:sqlite';
-import { createLongHorizonAgentTables } from '../../../src/storage/schema';
 import { createEvolutionTables } from '../../../src/storage/schema/evolution';
+import { createLongHorizonAgentTables } from '../../../src/storage/schema/long-horizon-agents';
 
 export function createSpaceTables(db: BunDatabase): void {
 	db.exec('PRAGMA foreign_keys = ON');
